refactor: move secret detection logic to utils and update related tests
